HW2.5/requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiofiles==25.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.3
async-lru==2.0.5
attrs==25.4.0
backports.zstd==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.42.15
botocore==1.42.15
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dacite==1.9.2
Deprecated==1.3.1
diffusers==0.36.0
dill==0.4.0
dnspython==2.8.0
exceptiongroup==1.3.1
fiftyone==1.11.0
fiftyone-brain==0.21.4
fiftyone_db==1.4.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
ftfy==6.3.1
glob2==0.7
graphql-core==3.2.7
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.15.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
inflate64==1.0.4
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kagglehub==0.3.13
kiwisolver==1.4.9
lazy_loader==0.4
lia-web==0.2.3
lightning-utilities==0.15.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.14.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
paramiko==3.5.1
pillow==12.0.0
plotly==6.5.0
polars==1.36.1
polars-runtime-32==1.36.1
pprintpp==0.4.0
priority==2.0.0
protobuf==6.33.2
psutil==7.1.3
py7zr==1.1.0
pybcj==1.0.7
pycocotools==2.0.11
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydash==8.0.5
pymongo==4.9.2
PyNaCl==1.6.1
pyparsing==3.2.5
pyppmd==1.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rarfile==4.2
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rtree==1.4.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.3
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.1
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.50.0
strawberry-graphql==0.287.3
stringzilla==4.5.1
sympy==1.14.0
tabulate==0.9.0
taskgroup==0.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.22
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.3.238
ultralytics-thop==2.0.18
universal-analytics-python3==1.1.1
urllib3==2.6.2
voxel51-eta==0.15.1
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==2.0.1
wsproto==1.3.2
xmltodict==1.0.2
zipp==3.23.0
zmq==0.0.0
